Introduces the Textfield class.
This is a more basic, primary widget that can be used to draw more complex
ones.

By itself it does not expose much logic or rules; rather, it can be considered
a canvas to draw other widgets on. The old label has been updated to use this
as a component and instead inforce its own rules an interface upon the
Textfield.