Renames BorderPrototype to BorderStyle. Fixes errors with border destruction.
The new name more accuractly describes what this object is and does. In
addition, allocation and deallocation process for Borders has been improved; it
now properly removes old borders when going out of scope, as well as being able
to replace borders during runtime if and only if it is in a window that already
has allocated room for borders.

This, as a necessity, introduces the Border::set method. Some refactoring came
as a consequence, where the logic in a Border constructor was moved out into a
static method to be reused in the set method.