Introduces the Button class.
This is an object that associates some actions with a label. Actions include
appearance change when focus are toggled as well as a function called whenever
the button is triggered.

This requires Label::redraw to also call write on the underlying Textfield; it
can never be sure that nothing has changed.