Changed implementation of widget to use std unique_ptr instead of homegrown.
This is a move to simplify Widget's implementation a tad. This also removes the
need to explicitly implement all functions towards ncurses twice; once
externally and once internally. In addition this now uses a standardized and
probably safer implementation of a RAII-style unique pointer. Requires a custom
deleter, which is private within Widget.

Updated all Widget method calls to support this.

An added gain with this model is out-of-the-box movability; as unique_ptr
implements a move constructor, it automatically makes Widget a bit richer as it
opens for being moved as well.